requirements.txt
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
xmltodict==0.14.2
zipp==3.20.2
